refactor: use shared lib file extension references
add glob pattern generator and file extension glob pattern exports to the shared lib
replace hardcoded glob strings with generated imports
refactor duplicated file path creation code
remove hardcoded module extension reference
