+++ packages/shared/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/Shared';
+
+export * from './lib/fileExtensions';
